+++ backend/tests/services/AuthService.test.ts
@@ -1,3 +1,4 @@
+import jwt from 'jsonwebtoken';
 import AuthService from '../../src/services/AuthService';
 
 describe('Testing AuthService', () => {
@@ -18,9 +19,36 @@ describe('Testing AuthService', () => {
     expect(() => service.getToken('wrong')).toThrow(expect.objectContaining({ message: 'Invalid password' }));
   });
 
-  it('Requesting a token using a valid password', () => {
+  it.each([
+    [
+      'superadmin',
+      {
+        sub: 'super-admin',
+        scope: 'super-admin',
+        email_verified: true,
+        given_name: 'super',
+        family_name: 'admin',
+        name: 'super admin',
+        email: 'super-admin@localhost',
+      },
+    ],
+    [
+      'dataadmin',
+      {
+        sub: 'data-admin',
+        scope: 'data-admin',
+        email_verified: true,
+        given_name: 'data',
+        family_name: 'admin',
+        name: 'data admin',
+        email: 'data-admin@localhost',
+      },
+    ],
+  ])('Requesting a token using a valid password', (password: string, expectedTokenClaims: any) => {
     const service = new AuthService();
-    service.getToken('superadmin');
-    service.getToken('dataadmin');
+    const token = service.getToken(password);
+    const decoded = jwt.decode(token, { complete: true });
+    expect(decoded).toBeDefined();
+    expect(decoded!.payload).toMatchObject(expectedTokenClaims);
   });
 });