+++ frontend/soilhive-app/src/layouts/AdminPortalLayout/AdminPortalLayout.tsx
@@ -1,0 +1,19 @@
+import { Outlet } from 'react-router';
+import { AdminHeader } from 'components/AdminPortal/AdminHeader/AdminHeader';
+import { AdminSidebar } from 'components/AdminPortal/AdminSidebar/AdminSidebar';
+
+import styles from './AdminPortalLayout.module.scss';
+
+export function AdminPortalLayout() {
+  return (
+    <>
+      <AdminHeader></AdminHeader>
+      <div className={styles.Layout}>
+        <AdminSidebar />
+        <main className={styles.Content}>
+          <Outlet />
+        </main>
+      </div>
+    </>
+  );
+}