+++ frontend/soilhive-app/src/types/backend.ts
@@ -182,3 +182,13 @@ export interface RasterFilterCategory {
   updated_at: string;
   deleted_at: string | null;
 }
+
+export interface License {
+  id: string;
+  name: string;
+  slug?: string;
+  full_name?: string;
+  url?: string;
+  created_at: Date;
+  updated_at: Date | null;
+}